test: add search test case
